pyproject.toml
[tool.poetry]
name = "flow-stock"
version = "0.1.34"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "==3.10.*"
httpx = "^0.28.1"
pycryptodome = "^3.21.0"
kafka-python = "^2.0.2"
requests = "^2.32.3"
prefect = "2.18.3"
websockets = "==12.*"
griffe = "==0.*"
python-dotenv = "^1.0.1"
pytz = "^2024.2"
pytest = "^8.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

